整理meta tag相關程式碼
